renaming redirectToast to Toast
